requirements.txt
streamlit
streamlit-folium
folium
google-generativeai
requests
Pillow
upstage